Pull tfprovider code into package. Move to ginkgo.
