+++ tfprovider/tfprovider_suite_test.go
@@ -1,0 +1,13 @@
+package tfprovider_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestTfprovider(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Tfprovider Suite")
+}
